+++ packages/dataviews/src/dataform-controls/select.tsx
@@ -1,14 +1,17 @@
 /**
  * WordPress dependencies
  */
-import { SelectControl } from '@wordpress/components';
-import { useCallback } from '@wordpress/element';
+import { privateApis } from '@wordpress/components';
+import { useCallback, useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import type { DataFormControlProps } from '../types';
+import { unlock } from '../lock-unlock';
+
+const { ValidatedSelectControl } = unlock( privateApis );
 
 export default function Select< Item >( {
 	data,
@@ -16,7 +19,14 @@ export default function Select< Item >( {
 	onChange,
 	hideLabelFromVision,
 }: DataFormControlProps< Item > ) {
-	const { id, label, type } = field;
+	const { id, type, label, description } = field;
+	const [ customValidity, setCustomValidity ] =
+		useState<
+			React.ComponentProps<
+				typeof ValidatedSelectControl
+			>[ 'customValidity' ]
+		>( undefined );
+
 	const isMultiple = type === 'array';
 	const value = field.getValue( { item: data } ) ?? ( isMultiple ? [] : '' );
 	const onChangeControl = useCallback(
@@ -48,10 +58,31 @@ export default function Select< Item >( {
 			  ];
 
 	return (
-		<SelectControl
+		<ValidatedSelectControl
+			required={ !! field.isValid?.required }
+			onValidate={ ( newValue: any ) => {
+				const message = field.isValid?.custom?.(
+					{
+						...data,
+						[ id ]: newValue,
+					},
+					field
+				);
+
+				if ( message ) {
+					setCustomValidity( {
+						type: 'invalid',
+						message,
+					} );
+					return;
+				}
+
+				setCustomValidity( undefined );
+			} }
+			customValidity={ customValidity }
 			label={ label }
 			value={ value }
-			help={ field.description }
+			help={ description }
 			options={ elements }
 			onChange={ onChangeControl }
 			__next40pxDefaultSize